# Docs/tickets_openapi.yaml
openapi: 3.0.3
info:
  title: Tickets Service API
  version: '1.0'
  description: |
    Railway ticket service implemented as a JAX-RS application deployed as a WAR to WildFly.

    The service owns tickets, temporary departure venues and price histories. Inventory
    train-set data is consumed from the Redpanda CDC topic and stored as JSONB snapshots;
    train sets, train-set carriages and seats are read-only projections of those snapshots.
servers:
  - url: http://localhost:4455/api/v1/tickets
    description: Oathkeeper gateway
  - url: http://localhost:8074/api/v1/tickets
    description: Direct WildFly application endpoint (Docker Compose host port)
...
paths:
...
  /price-histories:
    get:
      tags: [Price histories]
      summary: List price-history records
      description: Returns all historical price states with pagination, sorting and filtering.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
        - name: id
          in: query
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: ticketId
          in: query
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: changedDateFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: changedDateTo
          in: query
          schema:
            type: string
            format: date-time
        - name: basePrice
          in: query
          schema:
            type: number
            format: double
            minimum: 0
            exclusiveMinimum: true
        - name: discount
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Price-history records returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceHistoryPageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [Price histories]
      summary: Append a price-history record
      description: Adds a new current price/discount state for an existing ticket. The history is append-only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriceHistoryCreateRequest'
      responses:
        '201':
          description: Price-history record created successfully.
          headers:
            Location:
              description: URL of the created history record.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceHistory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /price-histories/{id}:
    get:
      tags: [Price histories]
      summary: Get price-history record by ID
      parameters:
        - $ref: '#/components/parameters/PriceHistoryId'
      responses:
        '200':
          description: Price-history record found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceHistory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /train-sets:
    get:
      tags: [Inventory projection]
      summary: List train-set projections
      description: |
        Returns the latest available snapshot for each Inventory train_set_id. No relational
        train_sets table is maintained in Tickets; data is projected from train_set_cdc.data JSONB.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
        - name: id
          in: query
          schema:
            type: integer
            format: int32
            minimum: 1
        - name: code
          in: query
          schema:
            type: string
        - name: name
          in: query
          schema:
            type: string
        - name: buildNumber
          in: query
          schema:
            type: integer
            minimum: 1
        - name: technicalName
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Train-set projections returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainSetPageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /train-sets/{trainSetId}:
    get:
      tags: [Inventory projection]
      summary: Get train-set projection
      parameters:
        - $ref: '#/components/parameters/TrainSetId'
      responses:
        '200':
          description: Train-set projection found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainSet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /train-sets/{trainSetId}/carriages:
    get:
      tags: [Inventory projection]
      summary: List carriages in a train set
      parameters:
        - $ref: '#/components/parameters/TrainSetId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: sort
          in: query
          description: Repeatable sort expression; default is position,asc.
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          example: [position,asc]
        - name: id
          in: query
          schema:
            type: integer
            format: int32
            minimum: 1
        - name: carriageNumber
          in: query
          schema:
            type: string
            maxLength: 16
        - name: position
          in: query
          schema:
            type: integer
            minimum: 1
        - name: inventoryNumber
          in: query
          schema:
            type: string
        - name: serialNumber
          in: query
          schema:
            type: string
        - name: carriageTypeCode
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Carriages returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarriagePageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /train-sets/{trainSetId}/carriages/{carriageNumber}/seats:
    get:
      tags: [Inventory projection]
      summary: List seats of a train-set carriage
      parameters:
        - $ref: '#/components/parameters/TrainSetId'
        - name: carriageNumber
          in: path
          required: true
          schema:
            type: string
            maxLength: 16
        - $ref: '#/components/parameters/Page'
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sort
          in: query
          description: Repeatable sort expression; default is seatNumber,asc.
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          example: [seatNumber,asc]
        - name: seatNumber
          in: query
          schema:
            type: string
            maxLength: 3
        - name: x
          in: query
          schema:
            type: number
            format: double
        - name: y
          in: query
          schema:
            type: number
            format: double
        - name: rotation
          in: query
          schema:
            type: number
            format: double
      responses:
        '200':
          description: Seats returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatPageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
    Page:
      name: page
      in: query
      description: One-based page number.
      schema:
        type: integer
        minimum: 1
        default: 1
    Size:
      name: size
      in: query
      description: Page size; maximum 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Sort:
      name: sort
      in: query
      description: Repeatable sort expression in the form field,direction where direction is asc or desc.
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
      example: [creationDate,desc]
...
    PriceHistoryId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    TrainSetId:
      name: trainSetId
      in: path
      required: true
      schema:
        type: integer
        format: int32
        minimum: 1
  responses:
    BadRequest:
      description: Request validation or parameter error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested object was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    PriceHistory:
      type: object
      required:
        - id
        - ticketId
        - changedDate
        - discount
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          readOnly: true
        ticketId:
          type: integer
          format: int64
          minimum: 1
          readOnly: true
        changedDate:
          type: string
          format: date-time
          readOnly: true
        basePrice:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true
          nullable: true
        discount:
          type: integer
          minimum: 1
          maximum: 100
    PriceHistoryCreateRequest:
      type: object
      required:
        - ticketId
        - discount
      properties:
        ticketId:
          type: integer
          format: int64
          minimum: 1
        basePrice:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true
          nullable: true
        discount:
          type: integer
          minimum: 1
          maximum: 100
    TrainSet:
      type: object
      required:
        - id
        - code
        - name
        - buildNumber
        - technicalName
        - snapshotVersion
      properties:
        id:
          type: integer
          format: int32
          minimum: 1
        code:
          type: string
        name:
          type: string
        buildNumber:
          type: integer
          minimum: 1
        technicalName:
          type: string
        description:
          type: string
          nullable: true
        snapshotVersion:
          type: integer
          format: int64
          minimum: 1
          description: Version from inventory.train_set_cdc.
    CarriageType:
      type: object
      required: [id, code, name]
      properties:
        id:
          type: integer
          format: int32
        code:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
    Scheme:
      type: object
      required: [id, code, name, version, storageKey]
      properties:
        id:
          type: integer
          format: int32
        code:
          type: string
        name:
          type: string
        version:
          type: integer
          minimum: 1
        storageKey:
          type: string
    Carriage:
      type: object
      required:
        - id
        - carriageNumber
        - position
        - carriageType
        - serialNumber
        - inventoryNumber
        - scheme
      properties:
        id:
          type: integer
          format: int32
        carriageNumber:
          type: string
          maxLength: 16
        position:
          type: integer
          minimum: 1
        carriageType:
          $ref: '#/components/schemas/CarriageType'
        serialNumber:
          type: string
        inventoryNumber:
          type: string
        scheme:
          $ref: '#/components/schemas/Scheme'
    Seat:
      type: object
      required:
        - id
        - seatNumber
        - x
        - y
        - rotation
      properties:
        id:
          type: integer
          format: int32
        seatNumber:
          type: string
          minLength: 1
          maxLength: 3
        x:
          type: number
          format: double
        y:
          type: number
          format: double
        rotation:
          type: number
          format: double
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - timestamp
        - path
      properties:
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
...
    PriceHistoryPageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages, hasNext, hasPrevious]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PriceHistory'
        page:
          type: integer
          minimum: 1
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
    TrainSetPageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages, hasNext, hasPrevious]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/TrainSet'
        page:
          type: integer
          minimum: 1
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
    CarriagePageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages, hasNext, hasPrevious]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Carriage'
        page:
          type: integer
          minimum: 1
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
    SeatPageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages, hasNext, hasPrevious]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Seat'
        page:
          type: integer
          minimum: 1
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean