# Services/tickets/src/main/java/com/fuzis/tickets/dto/PriceHistoryCreateRequest.java
package com.fuzis.tickets.dto;

import jakarta.validation.constraints.Max;
import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Positive;

import java.math.BigDecimal;

public class PriceHistoryCreateRequest {
    @NotNull @Positive
    private Long ticketId;
    @Positive
    private BigDecimal basePrice;
    @NotNull @Min(1) @Max(100)
    private Integer discount;
    public Long getTicketId() { return ticketId; }
    public void setTicketId(Long ticketId) { this.ticketId = ticketId; }
    public BigDecimal getBasePrice() { return basePrice; }
    public void setBasePrice(BigDecimal basePrice) { this.basePrice = basePrice; }
    public Integer getDiscount() { return discount; }
    public void setDiscount(Integer discount) { this.discount = discount; }
}